+++ Frontend/app/main/components/contacts/add-contact.component.ts
@@ -1,28 +1,36 @@
-import {Component, OnInit} from '@angular/core';
+import {Component, OnDestroy, OnInit} from '@angular/core';
 import {FormControl, FormGroup} from '@angular/forms';
 import {ActivatedRoute, Router} from '@angular/router';
-import {CRMService} from '../../services/crm.service';
+import {CRMDataService} from '../../services/crm-data.service';
 import {Contact} from '../../models/contact.model';
 import {ToastsManager} from 'ng2-toastr';
+import {CRMStoreService} from '../../services/crm-store.service';
+import {Subscription} from 'rxjs/Subscription';
 @Component({
 	selector: 'create-contact-component',
 	template: `
-		<h1>Add Contact</h1>
-		<form [formGroup]="addForm">
-			<single-line-text-input-component label="Name" [(model)]="newContact.name"
-											  [control]="nameControl"></single-line-text-input-component>
-			<single-line-text-input-component label="Phone" [(model)]="newContact.phone"
-											  [control]="phoneControl"></single-line-text-input-component>
-			<single-line-text-input-component label="Email" [(model)]="newContact.email"
-											  [control]="emailControl"></single-line-text-input-component>
-			<single-line-text-input-component label="Position" [(model)]="newContact.position"
-											  [control]="positionControl"></single-line-text-input-component>
-			<button type="reset" class="btn-warning btn-lg pull-right" [routerLink]="['/Contacts/company_id']">Cancel</button>
-			<button type="button" class="btn-success btn-lg pull-right" (click)="onSubmit(addForm.value)">Submit</button>
-		</form>
+		<h4>Add Contact</h4>
+		<div *ngIf="!ownerID">
+			<h5>Please Select A Company First</h5>
+		</div>
+		<div *ngIf="!!ownerID">
+			<form [formGroup]="addForm">
+				<single-line-text-input-component label="Name" [(model)]="newContact.name"
+												  [control]="nameControl"></single-line-text-input-component>
+				<single-line-text-input-component label="Phone" [(model)]="newContact.phone"
+												  [control]="phoneControl"></single-line-text-input-component>
+				<single-line-text-input-component label="Email" [(model)]="newContact.email"
+												  [control]="emailControl"></single-line-text-input-component>
+				<single-line-text-input-component label="Position" [(model)]="newContact.position"
+												  [control]="positionControl"></single-line-text-input-component>
+				<button type="reset" class="btn-warning btn-lg pull-right" [routerLink]="['/Contacts']">Cancel</button>
+				<button type="button" class="btn-success btn-lg pull-right" (click)="onSubmit(addForm.value)">Submit</button>
+			</form>
+		</div>
 	`
 })
-export class AddContactComponent implements OnInit {
+export class AddContactComponent implements OnInit, OnDestroy {
+	public crmStateSub: Subscription;
 	public newContact: Contact = <Contact>{};
 	public ownerID: string;
 	public nameControl: FormControl = new FormControl('', []);
@@ -38,21 +46,29 @@ export class AddContactComponent implements OnInit {
 	constructor(
 		public toastr: ToastsManager,
 		private router: Router,
-		private crmService: CRMService,
-		private route: ActivatedRoute,
+		private crmData: CRMDataService,
+		private crmStore: CRMStoreService,
 	){}
 
+	public ngOnDestroy(): void {
+		this.crmStateSub.unsubscribe();
+	}
+
 	public ngOnInit(): void {
-		this.route.params.subscribe(params => {
-			this.ownerID = params.owner_id;
-		})
+			this.crmStateSub = this.crmStore.crmStore$.subscribe(state => {
+				if (state.selectedCompany && state.selectedCompany.id){
+					this.ownerID = state.selectedCompany.id;
+				} else {
+					this.ownerID = null;
+				}
+			});
 	}
 
 	public onSubmit(values): void {
 			console.log('submitting contact', values);
-			this.crmService.newContact({owner_id: this.ownerID, props: values}).then(res => {
+			this.crmData.newContact({owner_id: this.ownerID, props: values}).then(res => {
 			this.toastr.success(res.name + ' added!');
-			this.router.navigate(['/Companies']);
+			this.router.navigate(['/Contacts']);
 		}).catch(err => {
 			this.toastr.error('Oh No! Error with adding contact.');
 			console.log(err)