+++ Frontend/app/main/components/navbar.component.ts
@@ -1,6 +1,7 @@
 import {Component} from '@angular/core';
+import {CRMStoreService} from '../services/crm-store.service';
 @Component({
-	selector: 'navbar',
+	selector: 'navbar-component',
 	template: `
 	<div class="row">
 		<ul class="nav nav-tabs nav-tabs-justified">
@@ -8,10 +9,10 @@ import {Component} from '@angular/core';
 				<a [routerLink]="['/Companies']"><span class="icon icon-office"></span></a>
 			</li>
 			<li class="tab" role="tab">
-				<a [routerLink]="['/Contacts/All']" ><span class="icon icon-user-tie"></span></a>
+				<a [routerLink]="['/Contacts']" ><span class="icon icon-user-tie"></span></a>
 			</li>
 			<li class="tab" role="tab">
-				<a [routerLink]="['/Quotes/All']" ><span class="icon icon-bubble2"></span></a>
+				<a [routerLink]="['/Quotes']" ><span class="icon icon-bubble2"></span></a>
 			</li>
 			<li class="tab" role="tab">
 				<a [routerLink]="['/Settings']" ><span class="icon icon-cog"></span></a>
@@ -22,5 +23,4 @@ import {Component} from '@angular/core';
 
 })
 export class NavbarComponent {
-
 }